Overhaul key input system
There was an issue when moving from React - for a given key sequence, earlier
keybinds that set the UI mode would mean that keybinds after it that are for
the new UI mode would be triggered from the same keypress.

The thing that fixes that here is that the UI mode is read once, then
all keybinds are tested. Any update of the UI mode doesn't change any key
processing for the keypress that caused it.